requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
python-dotenv
bcrypt
python-jose[cryptography]
yookassa
httpx